pyproject.toml
[project]
name = "qa_framework"
version = "0.1.0"
description = "Data quality assurance framework for Databricks"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pyspark>=3.4.0",
    "databricks-sdk>=0.18.0",
    "pydantic>=2.0.0",
    "pandas>=2.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "great-expectations>=0.18.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.1",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["code"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=qa_framework",
    "--cov-report=term-missing",
]

[tool.coverage.run]
source = ["code"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
